Add standalone quartic validation figure
For ease of review, this is mostly self-contained.

For each of 40 combinations of {schemes} x {timesteps}, computes (A)
high-precision histograms of the steady-state distribution over x and
(x,v) in `quartic_kl_histograms.py`, and (B) corresponding
nonequilibrium estimates of the KL divergence
in`quartic_validation.py`. These are both plotted in
`quartic_kl_plot.py`. These can be run in parallel in a few minutes on
the cluster using `quartic_reproduce.sh`, but note that something must
be wrong with line 9, since it doesn't actually generate the plot until
you go back afterwards and execute: `bsub < quartic_kl_plot.lsf`...