Quick check whether low-order expansion is more feasible now
Previously we had abandoned this idea, because var(w) >> 1 for the
testsystems / timesteps of interest. However, we have since switched to
a more-manageable testsystem for validation, and are now more acutely
aware of the difficulty of controlling the bias of the nested monte
carlo scheme. Ideally, I would like to be able to compare the
near-equilibrium scheme to upper- and lower-bounds on the D_KL in
Figure 6 -- it is unsatisfying to compare the scheme only to a
lower-bound...